Fix sign-compare warnings
This fixes the following warning:

warning: comparison between signed and unsigned integer expressions
[-Wsign-compare]

Tested on Linux with the following command line:

$ bazel test :all

+++ re2/testing/exhaustive3_test.cc
@@ -84,7 +84,7 @@ TEST(InterestingUTF8, AB) {
     "[[:upper:]] [[:xdigit:]] [\\s\\S] [\\d\\D] [^\\w\\W] [^\\d\\D]");
   vector<string> ops;  // no ops
   vector<string> alpha = InterestingUTF8();
-  for (int i = 0; i < alpha.size(); i++)
+  for (size_t i = 0; i < alpha.size(); i++)
     alpha[i] = "a" + alpha[i] + "b";
   ExhaustiveTest(1, 0, atoms, ops,
                  1, alpha, "a%sb", "");
